// prisma/generated/zod/inputTypeSchemas/BookingRequestFlowUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableIntFieldUpdateOperationsInputSchema } from './NullableIntFieldUpdateOperationsInputSchema';
import { IntFieldUpdateOperationsInputSchema } from './IntFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { NullableJsonNullValueInputSchema } from './NullableJsonNullValueInputSchema';
import { InputJsonValueSchema } from './InputJsonValueSchema';
import { BoolFieldUpdateOperationsInputSchema } from './BoolFieldUpdateOperationsInputSchema';
import { NullableDateTimeFieldUpdateOperationsInputSchema } from './NullableDateTimeFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { BookingRequestUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema } from './BookingRequestUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema';
import { MerchantUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema } from './MerchantUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema';
import { UiTemplateUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema } from './UiTemplateUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema';

export const BookingRequestFlowUpdateInputSchema: z.ZodType<Prisma.BookingRequestFlowUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  slug: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  usageLimit: z.union([ z.number().int(),z.lazy(() => NullableIntFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  usageCount: z.union([ z.number().int(),z.lazy(() => IntFieldUpdateOperationsInputSchema) ]).optional(),
  webhookUrl: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  webhookEvents: z.union([ z.lazy(() => NullableJsonNullValueInputSchema),InputJsonValueSchema ]).optional(),
  isActive: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  expiresAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  bookingRequest: z.lazy(() => BookingRequestUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema).optional(),
  merchant: z.lazy(() => MerchantUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema).optional(),
  uiTemplate: z.lazy(() => UiTemplateUpdateOneRequiredWithoutBookingRequestFlowNestedInputSchema).optional()
}).strict();

export default BookingRequestFlowUpdateInputSchema;